requirements.txt
aiohttp
aiosignal
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
asgiref
async-timeout
attrs
autobahn
Automat
bcrypt
beautifulsoup4
bleach
cachetools
certifi
cffi
channels
channels-redis
charset-normalizer
colorama
constantly
contourpy
cryptography
cycler
daphne
dataclasses-json
defusedxml
Django
django-cleanup
django-environ
django-extra-views
django-redis
django-searchable-encrypted-fields
django-storages
djangorestframework
exceptiongroup
fonttools
frozenlist
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
gunicorn
hyperlink
idna
importlib-resources
incremental
joblib
jsonpatch
jsonpointer
kiwisolver
langchain
langsmith
marshmallow
matplotlib
multidict
mypy-extensions
mysqlclient
numpy
oauthlib
openai
packaging
pandas
Pillow
plotly
protobuf
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
pyOpenSSL
pyparsing
python-dateutil
python3-openid
pytz
PyYAML
redis
regex
requests
requests-oauthlib
rsa
scikit-learn
scipy
service-identity
six
sniffio
sorl-thumbnail
soupsieve
SQLAlchemy
sqlparse
stripe
tenacity
threadpoolctl
tiktoken
tinycss2
tqdm
Twisted
txaio
typing-inspect
typing_extensions
tzdata
Unidecode
urllib3
webencodings
wordcloud
yarl
zipp
zope.interface